Chunk merged signals into separate messages
One read may return multiple signals. Convert each signal into an
individual message.

The code does not handle partial signals. If partial reads are possible,
the gen_server will need to maintain the partial data in its state.

Running the tests, the VM is sometimes printing out the signal. The
behaviour of signals is mysterious. Presumably the VM has its own signal
handler that is also firing. Whether this will affect the state of the
VM is an open isssue.