getumask/0: safer umask for Linux
getumask/0 retrieves the process mask by reading /proc/self/status.
Unlike umask/0, getumask/0 does not affect the beam process state.

On other platforms or if the proc file is not parsable, getumask/0 will
fail back to using umask/0. This behaviour is not safe but conforms to
the umask(2) interface.

Callers can avoid umask/0 completely by using status/0 to retrieve the
process information as a map.